[mkt tst] wait for cyclic broadcasts when testing events order
